sprint1/step2 refactor
