--- src/service/burger-ingredients.service.ts
@@ -1,18 +1,0 @@
- class ingredientService {
-  private baseUrl: string;
-
-  constructor() {
-    this.baseUrl = 'https://norma.nomoreparties.space/api/ingredients';
-  }
- async getIngredients(){
-    const response = await fetch(this.baseUrl, {
-        method: 'GET',
-      });
-      const data = await response.json();
-
-      console.log(data?.data)
-      return data; 
-  }
-}
-
-export default new ingredientService();
